+++ packages/spotlight/CHANGELOG.md
@@ -6,6 +6,8 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Added
 
+- `isMuted()` method to `@enact/spotlight`
+- `spotlightMuted` prop to `@enact/spotlight/container`
 - `isPaused()`, `isSpottable()`, and `getCurrent()` methods
 
 ## [1.0.0-alpha.5] - 2016-12-16