+++ packages/spotlight/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `isMuted()` method to `@enact/spotlight`
+- `spotlightMuted` prop to `@enact/spotlight/container`
+- `isPaused()`, `isSpottable()`, and `getCurrent()` methods
+
+## [1.0.0-alpha.5] - 2016-12-16
+
+No changes.
+
 ## [1.0.0-alpha.4] - 2016-12-2
 
 ### Added